add parseInt

+++ js/index.js
@@ -1,4 +1,4 @@
-const userNumber = parseFloat(prompt('What position of Fibonacci sequence to find? Please, enter a number:')); 
+const userNumber = parseInt(prompt('What position of Fibonacci sequence to find? Please, enter a number:')); 
 const fibonacciSequence = [1, 1];
 while (userNumber > 0) {
     const sumOfNeighboringNumbers = fibonacciSequence[fibonacciSequence.length-1] + fibonacciSequence[fibonacciSequence.length-2];
@@ -7,3 +7,10 @@ while (userNumber > 0) {
         break;
 }
 alert('Your number is ' + fibonacciSequence[fibonacciSequence.length-1]);
+
+
+
+
+
+
+
